Don’t set alt to figcaption value by default
The alt and figcaption should not contain the same value as this would
duplicate the text for screen readers. If the alt is passed it will be
set, otherwise the result should be alt="". This does not enforce that
at least one of the alt or caption attributes is passed.