Avoid creating links with no discernable name
See 719053d28d73233a163f6fb4a8a8d8c9e843df73 for previous attempt. Whilst it is
better to avoid screen readers repeating the alt and caption texts, if the alt
attribute is empty and the figure has a URL, then the link does not have any
suitable link text (see https://web.dev/link-name/).

One solution would be to wrap the entire <figure> element in the link, but
the markdown engines do not allow an <a> to be block-level, i.e. to contain
block-level elements.