Create environment automatically again
Test that the `workflow_dispatch` event will create a deployment against
the branch that received the event. This means the manual steps can be
removed. This relies on Netlify using the commit ref in the dispatch,
but the field isn't available so we grab it from the API in the
Cloudflare worker.

See
https://answers.netlify.com/t/deploy-notification-payload-missing-commit-ref/155594