Manual environment creation again, more notification states
The magic environment creation doesn't work with forks because the ref
at Netlify is the name of the ref on the fork. It's not possible to call
a workflow_dispatch on a PR ref, so the environment needs to be manually
created.

Also subscribe to deploy preview started, failed and succeeded events,
not just succeeded.